+++ CHANGELOG.md
@@ -16,3 +16,5 @@ Please, use the format:
 ## [Unreleased]
 
 - beta-builder: automatically publish beta packages into pypi from main branch
+- improve debug logging
+- fix issues encountered with PR #11